requirements.txt
google-api-core==1.14.3
google-auth==1.6.3
google-auth-oauthlib==0.4.1
google-cloud-bigquery==1.21.0
google-cloud-bigquery-datatransfer==0.4.1
google-cloud-bigquery-storage==0.7.0
google-cloud-core==1.0.3
google-pasta==0.2.0
google-resumable-media==0.4.1
googleapis-common-protos==1.6.0
h5py==2.10.0
joblib==0.14.0
nltk==3.4.5
numpy==1.19.0
oauthlib==3.1.0
pandas==0.25.2
pandas-gbq==0.11.0
pandocfilters==1.4.2
prettytable==0.7.2
pyarrow==0.15.0
PyYAML==3.13
requests==2.23.0
requests-oauthlib==1.2.0
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.2.0
scikit-learn==0.21.3
scipy==1.4.1
six==1.14.0
sklearn==0.0
tqdm==4.37.0
urllib3==1.25.8
widgetsnbextension==3.5.1
wrapt==1.12.1